requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aiohttp==3.9.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
galileo==1.5.1
galileo-core
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.44
markupsafe==3.0.2
openai==1.84.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pydantic
pydantic-core
pydantic-settings
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
rich
flask
flask-cors
aiohttp
